NEW/requirements_fix.txt
# Django Core
Django==5.2.3
djangorestframework==3.15.2

# CORS and Security
django-cors-headers==4.5.0
whitenoise==6.8.2

# Environment and Configuration
python-decouple==3.8
dj-database-url==2.2.0

# Database
psycopg2-binary==2.9.11

# HTTP Requests
requests==2.32.3

# Rate Limiting
django-ratelimit==4.1.0

# Cache
redis==5.2.0

# Production Server
gunicorn==23.0.0

# Build Dependencies
setuptools==75.6.0
wheel==0.45.1